more weaponselectiontimeout msgs
git-svn-id: https://openlierox.svn.sourceforge.net/svnroot/openlierox@3537 34602234-ff1f-0410-a465-ea8f3b77ab7f

+++ src/client/CClient_Game.cpp
@@ -1012,7 +1012,7 @@ void CClient::NewNet_DoLocalShot( CWorm *w )
 		speed = NormalizeVector( &vel );
 	}
 	
-	shot.nAngle = Angle;
+	shot.nAngle = (int)Angle;
 	shot.nSpeed = (int)( speed*100 );
